refactor : Topic 관련 클래스들 패키지 변경
- TopicByDate의 성격상 엔티티가 아닌 DTO가 적합해 DTO로 패키지 이동
- TopicManager의 셩격상 엔티티가 아닌 utils 패키지가 적합해 utils로 이동
